https://github.com/nasa/nos3/issues/422 - Refactor unit test stubs.  Add additional unit test asserts.
